+++ semcore/modal/src/Modal.jsx
@@ -18,6 +18,10 @@ import { cssVariableEnhance } from '@semcore/utils/lib/useCssVariable';
 import { useFocusLock } from '@semcore/utils/lib/use/useFocusLock';
 import { useContextTheme } from '@semcore/utils/lib/ThemeProvider';
 import { useColorResolver } from '@semcore/utils/lib/use/useColorResolver';
+import {
+  ZIndexStackingContextProvider,
+  useZIndexStacking,
+} from '@semcore/utils/lib/zIndexStacking';
 
 class ModalRoot extends Component {
   static displayName = 'Modal';
@@ -139,10 +143,12 @@ function Window(props) {
       duration={duration}
       ref={windowRef}
     >
-      <PortalProvider value={windowRef}>
-        {closable && <Modal.Close />}
-        <Children />
-      </PortalProvider>
+      <ZIndexStackingContextProvider designToken='z-index-modal'>
+        <PortalProvider value={windowRef}>
+          {closable && <Modal.Close />}
+          <Children />
+        </PortalProvider>
+      </ZIndexStackingContextProvider>
     </SWindow>,
   );
 }
@@ -153,9 +159,10 @@ function Overlay(props) {
   const overlayRef = React.useRef(null);
   usePreventScroll(visible, props.disablePreventScroll);
   useContextTheme(overlayRef, visible);
+  const zIndex = useZIndexStacking('z-index-modal');
 
   return sstyled(styles)(
-    <SOverlay render={FadeInOut} ref={overlayRef}>
+    <SOverlay render={FadeInOut} ref={overlayRef} zIndex={zIndex}>
       <OutsideClick root={overlayRef} onOutsideClick={onOutsideClick}>
         <Children />
       </OutsideClick>