+++ semcore/data-table/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [4.41.0] - 2024-07-26
+
+### Changed
+
+- Version minor update due to children dependencies update (`@semcore/scroll-area` [5.33.1 ~> 5.34.0], `@semcore/utils` [4.31.0 ~> 4.32.0], `@semcore/core` [2.28.0 ~> 2.29.0]).
+
 ## [4.40.10] - 2024-07-19
 
 ### Changed